requirements.txt
openai>=1.50.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
pydantic>=2.0.0
streamlit>=1.30.0
